+++ Mimir.Worker/Services/IItemProductCalculationService.cs
@@ -7,7 +7,7 @@ namespace Mimir.Worker.Services;
 
 public interface IItemProductCalculationService
 {
-    Task<int?> CalculateCombatPointAsync(ItemProduct itemProduct);
+    Task<long?> CalculateCombatPointAsync(ItemProduct itemProduct);
 
     Task<(int? crystal, int? crystalPerPrice)> CalculateCrystalMetricsAsync(ItemProduct itemProduct);
 }
